chore: enhance Azure Static Web Apps workflow with Node.js setup and pnpm caching; add new API handling logic
