+++ .github/workflows/azure-static-web-apps-lively-coast-0ae46e91e.yml
@@ -18,6 +18,30 @@ jobs:
       - name: Checkout GitHub Action
         uses: actions/checkout@v4
 
+      - name: Setup Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: '20'
+
+      - name: Install pnpm
+        uses: pnpm/action-setup@v2
+        with:
+          version: 9
+          run_install: false
+
+      - name: Get pnpm store directory
+        shell: bash
+        run: |
+          echo "STORE_PATH=$(pnpm store path --silent)" >> $GITHUB_ENV
+
+      - name: Setup pnpm cache
+        uses: actions/cache@v4
+        with:
+          path: ${{ env.STORE_PATH }}
+          key: ${{ runner.os }}-pnpm-store-${{ hashFiles('**/pnpm-lock.yaml') }}
+          restore-keys: |
+            ${{ runner.os }}-pnpm-store-
+
       - name: Set environment variables
         run: |
           PR_ENV_NAME=$(echo ${{ github.event.pull_request.head.ref }} | tr -cd '[:alnum:]' | tr '[:upper:]' '[:lower:]')
@@ -34,7 +58,6 @@ jobs:
           app_location: '/'
           api_location: 'api'
           output_location: 'out'
-          skip_deploy_on_missing_secrets: true
         env:
           NEXT_PUBLIC_LANGGRAPH_API_URL: ${{ env.NEXT_PUBLIC_LANGGRAPH_API_URL }}
           LANGSMITH_API_KEY: ${{ secrets.LANGSMITH_API_KEY }}
